using JSON, API's, and SQL server
In this project geodump.py reads the where.txt file. In the file there a list of locations nearby in which I have visited. The geodump.py proceeds to use Google's API to find the exact location of the destinations. I then proceed to use JSON to read and clean up the data given, with that cleaned up data I store it into a SQL file (geodata.sqlite)